+++ src/id_translation/dio/_data_structure_io.py
@@ -2,22 +2,21 @@
 
 from abc import abstractmethod
 from collections.abc import Sequence
-from typing import Any
+from typing import Any, Generic
 
 from ..offline import TranslationMap
-from ..types import IdType, NameType, SourceType
+from ..types import IdType, NameType, SourceType, TranslatableT
 
 
-class DataStructureIO:
+class DataStructureIO(Generic[TranslatableT, NameType, SourceType, IdType]):
     """Insertion and extraction of IDs and translations."""
 
-    @staticmethod
+    @classmethod
     @abstractmethod
-    def handles_type(arg: Any) -> bool:
+    def handles_type(cls, arg: Any) -> bool:
         """Return ``True`` if the implementation handles data for the type of `arg`."""
 
-    @staticmethod
-    def names(translatable: Any) -> list[NameType] | None:
+    def names(self, translatable: TranslatableT) -> list[NameType] | None:
         """Extract names from `translatable`.
 
         Args:
@@ -28,9 +27,8 @@ class DataStructureIO:
         """
         return translatable.name if hasattr(translatable, "name") else None
 
-    @staticmethod
     @abstractmethod
-    def extract(translatable: Any, names: list[NameType]) -> dict[NameType, Sequence[IdType]]:
+    def extract(self, translatable: TranslatableT, names: list[NameType]) -> dict[NameType, Sequence[IdType]]:
         """Extract IDs from `translatable`.
 
         Args:
@@ -41,10 +39,13 @@ class DataStructureIO:
             A dict ``{name: ids}``.
         """
 
-    @staticmethod
     @abstractmethod
     def insert(
-        translatable: Any, names: list[NameType], tmap: TranslationMap[NameType, SourceType, IdType], copy: bool
+        self,
+        translatable: TranslatableT,  # TODO Higher-Kinded TypeVars
+        names: list[NameType],
+        tmap: TranslationMap[NameType, SourceType, IdType],
+        copy: bool,
     ) -> Any | None:
         """Insert translations into `translatable`.