+++ tests/dio/test_insert.py
@@ -1,19 +1,26 @@
+from collections.abc import Collection
+from typing import Any, Mapping, TypeAlias
+
 import numpy as np
 import pandas as pd
 import pytest
-from id_translation.dio import resolve_io
+from id_translation.dio import DataStructureIO, resolve_io
 from id_translation.dio.exceptions import NotInplaceTranslatableError
+from id_translation.types import IdType, SourceType
 
 from .conftest import TRANSLATED, UNTRANSLATED
 
 NAME = "nconst"
 
+CollectionOrStrDict = dict[str, IdType] | Mapping[str, IdType] | Collection[IdType]
+DIO: TypeAlias = DataStructureIO[CollectionOrStrDict[IdType], str, SourceType, IdType]
+
 
 def test_dict_insert(translation_map):
     expected = {"firstTitle": {TRANSLATED["firstTitle"][2]}, "nconst": TRANSLATED["nconst"].copy()}
-    dict_io = resolve_io(expected)
 
     actual = {"firstTitle": {UNTRANSLATED["firstTitle"][2]}, "nconst": UNTRANSLATED["nconst"].copy()}
+    dict_io: DIO[str, int] = resolve_io(actual)
     copied = dict_io.insert(actual, list(actual), translation_map, copy=True)
     assert copied == expected
 
@@ -31,7 +38,7 @@ def test_sequence_insert(ttype, translation_map):
 @pytest.mark.parametrize("ttype", [list, pd.Series, set])
 def test_sequence_insert_inplace(ttype, translation_map):
     actual = ttype(UNTRANSLATED[NAME])
-    translatable_io = resolve_io(actual)
+    translatable_io: DataStructureIO[Any, str, str, str] = resolve_io(actual)
     ans = translatable_io.insert(actual, [NAME], translation_map, copy=False)
     assert ans is None
     _test_eq(actual, ttype(TRANSLATED[NAME]))
@@ -39,7 +46,7 @@ def test_sequence_insert_inplace(ttype, translation_map):
 
 def test_insert_inplace_array(translation_map):
     actual = np.array(UNTRANSLATED[NAME], dtype=object)
-    translatable_io = resolve_io(actual)
+    translatable_io: DataStructureIO[Any, str, str, str] = resolve_io(actual)
     ans = translatable_io.insert(actual, [NAME], translation_map, copy=False)
     assert ans is None
     _test_eq(actual, np.array(TRANSLATED[NAME]))
@@ -47,7 +54,7 @@ def test_insert_inplace_array(translation_map):
 
 def test_forbidden_insert_inplace(translation_map):
     actual = tuple(UNTRANSLATED[NAME])
-    translatable_io = resolve_io(actual)
+    translatable_io: DIO[int, str] = resolve_io(actual)
 
     with pytest.raises(NotInplaceTranslatableError):
         translatable_io.insert(actual, [NAME], translation_map, copy=False)
@@ -69,9 +76,12 @@ def test_large_series(translation_map, monkeypatch):
     large_list = UNTRANSLATED[NAME] * 1000
     large_series = pd.Series(large_list)
 
-    resolve_io(large_list).insert(large_list, [NAME], translation_map, copy=False)
+    list_io: DIO[int, str] = resolve_io(large_list)
+    list_io.insert(large_list, [NAME], translation_map, copy=False)
     assert num_getitem_calls == len(large_list)  # Not actually needed; fewer would be nice!
-    resolve_io(large_series).insert(large_series, [NAME], translation_map, copy=False)
+
+    series_io: DIO[int, str] = resolve_io(large_series)
+    series_io.insert(large_series, [NAME], translation_map, copy=False)
     assert num_getitem_calls == len(large_list) + large_series.nunique()
 
     assert TRANSLATED[NAME] * 1000 == large_list
@@ -80,7 +90,7 @@ def test_large_series(translation_map, monkeypatch):
 
 def _do_insert(translation_map, ttype, copy):
     actual = ttype(UNTRANSLATED[NAME])
-    translatable_io = resolve_io(actual)
+    translatable_io: DIO[int, str] = resolve_io(actual)
     ans = translatable_io.insert(actual, [NAME], translation_map, copy=copy)
     return actual, ans