+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Consume `[transform]`-sections in auxiliary configuration files (#231).
+- Added typehints to `dio.DataStructureIO` and other `id_translation.dio` classes and functions.
 
 ### Fixed
 - Verify top-level sections in auxiliary configuration files.